refactor: separate model package, and app is selfish package
